Moved ProxyPassword to utils.ControllerCipher
I moved the tests under test_utils.py
